Upgraded SFML2. Minor changes for linux. Corrected EventsRenderingHelper::DrawTextInArea ( segfault on linux ).
git-svn-id: svn://localhost@512 8062f311-0dae-4547-b526-b8ab9ac864a5

+++ GDL/GDL/ManualTimer.h
@@ -23,7 +23,7 @@ class GD_API ManualTimer
 
         /**
          * Update the time of the timer
-         * @param Time to add in seconds
+         * @param Time to add in milliseconds
          */
         inline void UpdateTime(float time_) { if (!isPaused) time += time_; };
 
@@ -33,7 +33,7 @@ class GD_API ManualTimer
         inline void Reset() { time = 0; };
 
         /**
-         * Get the current time elapsed.
+         * Get the current time elapsed, in milliseconds.
          * @return Time elapsed
          */
         inline float GetTime() const { return time; };
@@ -59,7 +59,7 @@ class GD_API ManualTimer
     private:
 
         std::string name;
-        float time;
+        unsigned int time;
         bool isPaused;
 };
